build: move "nm-vpnc-service-defines.h" to "shared/nm-service-defines.h"
nm-openvpn has a similar file that serves the same purpose.

Instead of naming them different for each VPN plugin, give them all
the same name (which makes it more obvious how they relate, and it's
more convenient when looking at a diff between plugins).
