+++ auth-dialog/main.c
@@ -36,7 +36,7 @@
 #include <nma-vpn-password-dialog.h>
 #include <nm-vpn-service-plugin.h>
 
-#include "src/nm-ssh-service-defines.h"
+#include "nm-service-defines.h"
 
 #define KEYRING_UUID_TAG "connection-uuid"
 #define KEYRING_SN_TAG "setting-name"
