+++ shared/nm-service-defines.h
